+++ projects/packages/classic-theme-helper/src/class-social-links.php
@@ -31,14 +31,14 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 *
 		 * @var array
 		 */
-		private static $links;
+		private $links;
 
 		/**
 		 * A Publicize object.
 		 *
 		 * @var Publicize
 		 */
-		private static $publicize;
+		private $publicize;
 
 		/**
 		 * An array with all services that are supported by both Publicize and the
@@ -46,37 +46,23 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 *
 		 * @var array
 		 */
-		private static $services = array();
+		private $services = array();
 
 		/**
 		 * An array of the services the theme supports
 		 *
 		 * @var array
 		 */
-		private static $theme_supported_services = array();
+		private $theme_supported_services = array();
 
 		/**
-		 * Instantiate.
-		 *
-		 * All custom functionality will be hooked into the "init" action.
+		 * Constructor.
 		 */
-		public static function setup() {
-			add_action( 'init', array( __CLASS__, 'init' ), 30 );
-		}
-
-		/**
-		 * Conditionally hook into WordPress.
-		 *
-		 * Themes must declare that they support this module by adding
-		 * add_theme_support( 'social-links' ); during after_setup_theme.
-		 *
-		 * If no theme support is found there is no need to hook into WordPress. We'll
-		 * just return early instead.
-		 */
-		public static function init() {
+		public function __construct() {
 			if ( ! ( ! wp_is_block_theme() && current_theme_supports( 'social-links' ) && function_exists( 'publicize_init' ) ) ) {
 				return;
 			}
+
 			$theme_support = get_theme_support( 'social-links' );
 
 			/*
@@ -87,41 +73,41 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 				return;
 			}
 
-			self::$theme_supported_services = $theme_support[0];
-			self::$links                    = class_exists( Jetpack_Options::class ) ? Jetpack_Options::get_option( 'social_links', array() ) : '';
+			$this->theme_supported_services = $theme_support[0];
+			$this->links                    = class_exists( Jetpack_Options::class ) ? Jetpack_Options::get_option( 'social_links', array() ) : '';
 
-			self::admin_setup();
+			$this->admin_setup();
 
-			add_filter( 'jetpack_has_social_links', array( __CLASS__, 'has_social_links' ) );
-			add_filter( 'jetpack_get_social_links', array( __CLASS__, 'get_social_links' ) );
+			add_filter( 'jetpack_has_social_links', array( $this, 'has_social_links' ) );
+			add_filter( 'jetpack_get_social_links', array( $this, 'get_social_links' ) );
 
 			foreach ( $theme_support[0] as $service ) {
-				add_filter( "pre_option_jetpack-$service", array( __CLASS__, 'get_social_link_filter' ) ); // - `get_option( 'jetpack-service' );`
-				add_filter( "theme_mod_jetpack-$service", array( __CLASS__, 'get_social_link_filter' ) ); // - `get_theme_mod( 'jetpack-service' );`
+				add_filter( "pre_option_jetpack-$service", array( $this, 'get_social_link_filter' ) ); // - `get_option( 'jetpack-service' );`
+				add_filter( "theme_mod_jetpack-$service", array( $this, 'get_social_link_filter' ) ); // - `get_theme_mod( 'jetpack-service' );`
 			}
 		}
 
 		/**
 		 * Init the admin setup.
 		 */
-		public static function admin_setup() {
+		public function admin_setup() {
 			if ( ! current_user_can( 'manage_options' ) ) {
 				return;
 			}
 
-			if ( ! is_admin() && ! self::is_customize_preview() ) {
+			if ( ! is_admin() && ! $this->is_customize_preview() ) {
 				return;
 			}
 
-			// @phan-suppress-next-line PhanUndeclaredFunction -- See init() for the condition that ensures this function exists.
-			self::$publicize    = publicize_init();
-			$publicize_services = self::$publicize->get_services( 'connected' );
-			self::$services     = array_intersect( array_keys( $publicize_services ), self::$theme_supported_services );
+			// @phan-suppress-next-line PhanUndeclaredFunction -- See __construct() for the condition that ensures this function exists.
+			$this->publicize    = publicize_init();
+			$publicize_services = $this->publicize->get_services( 'connected' );
+			$this->services     = array_intersect( array_keys( $publicize_services ), $this->theme_supported_services );
 
-			add_action( 'publicize_connected', array( __CLASS__, 'check_links' ), 20 );
-			add_action( 'publicize_disconnected', array( __CLASS__, 'check_links' ), 20 );
-			add_action( 'customize_register', array( __CLASS__, 'customize_register' ) );
-			add_filter( 'sanitize_option_jetpack_options', array( __CLASS__, 'sanitize_link' ) );
+			add_action( 'publicize_connected', array( $this, 'check_links' ), 20 );
+			add_action( 'publicize_disconnected', array( $this, 'check_links' ), 20 );
+			add_action( 'customize_register', array( $this, 'customize_register' ) );
+			add_filter( 'sanitize_option_jetpack_options', array( $this, 'sanitize_link' ) );
 		}
 
 		/**
@@ -130,11 +116,11 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 *
 		 * @return void
 		 */
-		public static function check_links() {
-			$active_links = array_intersect_key( self::$links, array_flip( self::$services ) );
+		public function check_links() {
+			$active_links = array_intersect_key( $this->links, array_flip( $this->services ) );
 
-			if ( $active_links !== self::$links ) {
-				self::$links = $active_links;
+			if ( $active_links !== $this->links ) {
+				$this->links = $active_links;
 				if ( class_exists( Jetpack_Options::class ) ) {
 					Jetpack_Options::update_option( 'social_links', $active_links );
 				}
@@ -146,7 +132,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 *
 		 * @param WP_Customize_Manager $wp_customize Theme Customizer object.
 		 */
-		public static function customize_register( $wp_customize ) {
+		public function customize_register( $wp_customize ) {
 			$wp_customize->add_section(
 				'jetpack_social_links',
 				array(
@@ -156,8 +142,8 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 			);
 
 			if ( class_exists( Publicize::class ) ) {
-				foreach ( array_keys( self::$publicize->get_services( 'all' ) ) as $service ) {
-					$choices = self::get_customize_select( $service );
+				foreach ( array_keys( $this->publicize->get_services( 'all' ) ) as $service ) {
+					$choices = $this->get_customize_select( $service );
 
 					if ( empty( $choices ) ) {
 						continue;
@@ -174,7 +160,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 					$wp_customize->add_control(
 						"jetpack-$service",
 						array(
-							'label'    => self::$publicize->get_service_label( $service ),
+							'label'    => $this->publicize->get_service_label( $service ),
 							'section'  => 'jetpack_social_links',
 							'settings' => "jetpack_options[social_links][$service]",
 							'type'     => 'select',
@@ -191,8 +177,8 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 * @param array $option The incoming values to be sanitized.
 		 * @return array
 		 */
-		public static function sanitize_link( $option ) {
-			foreach ( self::$services as $service ) {
+		public function sanitize_link( $option ) {
+			foreach ( $this->services as $service ) {
 				if ( ! empty( $option['social_links'][ $service ] ) ) {
 					$option['social_links'][ $service ] = esc_url_raw( $option['social_links'][ $service ] );
 				} else {
@@ -208,8 +194,8 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 *
 		 * @return bool
 		 */
-		public static function has_social_links() {
-			return ! empty( self::$links );
+		public function has_social_links() {
+			return ! empty( $this->links );
 		}
 
 		/**
@@ -217,8 +203,8 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 *
 		 * @return array
 		 */
-		public static function get_social_links() {
-			return self::$links;
+		public function get_social_links() {
+			return $this->links;
 		}
 
 		/**
@@ -227,9 +213,9 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 * @param string $link The incoming value to be replaced.
 		 * @return string $link The social link that we've got.
 		 */
-		public static function get_social_link_filter( $link ) {
-			if ( preg_match( '/_jetpack-(.+)$/i', current_filter(), $matches ) && ! empty( self::$links[ $matches[1] ] ) ) {
-				return self::$links[ $matches[1] ];
+		public function get_social_link_filter( $link ) {
+			if ( preg_match( '/_jetpack-(.+)$/i', current_filter(), $matches ) && ! empty( $this->links[ $matches[1] ] ) ) {
+				return $this->links[ $matches[1] ];
 			}
 
 			return $link;
@@ -241,26 +227,26 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		 * @param string $service The social service.
 		 * @return array An associative array with profile links and display names.
 		 */
-		public static function get_customize_select( $service ) {
+		private function get_customize_select( $service ) {
 			$choices = array(
 				'' => __( '&mdash; Select &mdash;', 'jetpack-classic-theme-helper' ),
 			);
 
-			if ( isset( self::$links[ $service ] ) ) {
-				$choices[ self::$links[ $service ] ] = self::$links[ $service ];
+			if ( isset( $this->links[ $service ] ) ) {
+				$choices[ $this->links[ $service ] ] = $this->links[ $service ];
 			}
 
 			if ( class_exists( Publicize::class ) ) {
-				$connected_services = self::$publicize->get_services( 'connected' );
+				$connected_services = $this->publicize->get_services( 'connected' );
 				if ( isset( $connected_services[ $service ] ) ) {
 					foreach ( $connected_services[ $service ] as $c ) {
-						$profile_link = self::$publicize->get_profile_link( $service, $c );
+						$profile_link = $this->publicize->get_profile_link( $service, $c );
 
 						if ( false === $profile_link ) {
 							continue;
 						}
 
-						$choices[ $profile_link ] = self::$publicize->get_display_name( $service, $c );
+						$choices[ $profile_link ] = $this->publicize->get_display_name( $service, $c );
 					}
 				}
 			}
@@ -275,7 +261,7 @@ if ( ! class_exists( __NAMESPACE__ . '\Social_Links' ) ) {
 		/**
 		 * Back-compat function for versions prior to 4.0.
 		 */
-		public static function is_customize_preview() {
+		private function is_customize_preview() {
 			global $wp_customize;
 			return is_a( $wp_customize, 'WP_Customize_Manager' ) && $wp_customize->is_preview();
 		}