+++ projects/packages/my-jetpack/_inc/components/product-cards-section/protect-card/index.tsx
@@ -1,15 +1,12 @@
-import { type FC } from 'react';
+import { PRODUCT_SLUGS } from '../../../data/constants';
 import ProductCard from '../../connected-product-card';
 import ProtectValueSection from './protect-value-section';
+import type { ProductCardComponent } from '../types';
 
-const ProtectCard: FC< { admin: boolean; recommendation?: boolean } > = props => {
-	const slug = 'protect';
-
-	return (
-		<ProductCard { ...props } slug={ slug } upgradeInInterstitial={ true }>
-			<ProtectValueSection />
-		</ProductCard>
-	);
-};
+const ProtectCard: ProductCardComponent = props => (
+	<ProductCard { ...props } slug={ PRODUCT_SLUGS.PROTECT } upgradeInInterstitial>
+		<ProtectValueSection />
+	</ProductCard>
+);
 
 export default ProtectCard;