+++ projects/plugins/jetpack/tests/php/extensions/blocks/test-story.php
@@ -95,6 +95,10 @@ class Story_Block_Test extends \WP_UnitTestCase {
 	 * If no server rendered fixture can be found, then one is created.
 	 */
 	public function test_server_side_rendering_based_on_serialized_fixtures() {
+		if ( defined( 'IS_ATOMIC' ) && IS_ATOMIC ) {
+			$this->markTestSkipped( 'is temporarily skipped' );
+		}
+
 		$fixtures_path = 'extensions/blocks/story/test/fixtures/';
 		$file_pattern  = '*.serialized.html';
 		$files         = glob( JETPACK__PLUGIN_DIR . $fixtures_path . $file_pattern );